+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+### Changed
+### Fixed
+
+## [2.0.0] - 2022-10-27
+### Added
 - New operator `in_or_null`. Will resolve to `(col IN (values) OR col IS NULL)`. IF there are no values provided, it will resolve to `col IS NULL`.
 ### Changed
+- Upgrade to Ruby 3
 ### Fixed
 - Loading a persisted report will ignore any extra columns